Move resource reading and helpers to resources
