+++ indra/statements/concept.py
@@ -1,7 +1,11 @@
 from __future__ import absolute_import, print_function, unicode_literals
 from builtins import dict, str
 from future.utils import python_2_unicode_compatible
+import logging
 from collections import OrderedDict as _o
+
+
+logger = logging.getLogger(__name__)
 
 
 @python_2_unicode_compatible
